refactor: move notification slice to shared folder
