+++ src/contact/ContactForm.js
@@ -15,13 +15,16 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import React, { useEffect, useState } from 'react';
+
 import _ from 'lodash/fp';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
+import { addMessage } from 'terrasoApi/shared/notifications/notificationsSlice';
+
 import { useDocumentTitle } from 'common/document';
 import PageContainer from 'layout/PageContainer';
 import PageLoader from 'layout/PageLoader';
-import { addMessage } from 'notifications/notificationsSlice';
+
 import { HUBSPOT_FORMS } from 'config';
 import { useScript } from 'custom-hooks';
