+++ src/notifications/NotificationsWrapper.js
@@ -15,12 +15,14 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 import React, { useEffect } from 'react';
+
 import _ from 'lodash/fp';
 import { SnackbarProvider, useSnackbar } from 'notistack';
 import { Trans } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
+import { removeMessage } from 'terrasoApi/shared/notifications/notificationsSlice';
+
 import { Alert } from '@mui/material';
-import { removeMessage } from './notificationsSlice';
 
 const MAX_NOTIFICATIONS = 3;
 const AUTO_HIDE_DURATION = 10000;
